+++ AnEvoNet/protocol/anevonet_rpc.go
@@ -79,7 +79,7 @@ type Peer struct {
 }
 
 type Timestamp struct {
-	Sec int32 `thrift:"1,required" json:"sec"`
+	Sec int64 `thrift:"1,required" json:"sec"`
 }
 
 type ConnectionReq struct {
